ffs cant get rid of it printing garbage lengths
the way i understood it recvfrom returns either -1 or maybe even
another negative number or error.
Thats why i get printed some garbage length since i print it as a signed
integer.

So i decided to test the length for negatvitiy which seemed to work when
inspecting it with gdb. The sign flag gets set and the jump is being
followed. But the message somehow is still printed. This needs another
look but i rage afk for now.