Adapt insane assembly to be closer to what objdump disassembles
Makes debugging with gdb and objdump easier. To find this snippet in the
code base again.

nasm does some optimizations and does not use the full register when
we only store small values in it.

Also gdb and objdump represent values as hex not as decimal.
So if we also use hex in the code it makes searching the objdump/gdb
code in the source code easier.