gdb says this is fine strace says its bad
(gdb) info registers r8
r8             0x402013            4202515
(gdb) x/16b 0x402013
0x402013:       0       0       0       111     127     0       0
1
0x40201b:       0       0       0       0       0       0       0
0

this is from the 2nd send_udp call
where gdb clearly shows r8 is filled with a pointer to the sockaddr
correctly. And the content of that memory matches with what we put into
the bss section. ?!

but when ran with strace it shows the sendto call with some weird
arguments on the 2nd hit.

sendto(3,
"\4\0\0\377\377\377\377\5Q;YF\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0"...,
524, 0, {sa_family=AF_UNSPEC, sa_data="\0o\177\0\0\1\0\0\0\0\0\0\0\0"},
16) = 524

watfak is sa_data and it should not be sa_family=AF_UNSPEC