Do not use recursion to loop to avoid stack overflow
I did not reach any stack overflow. But while printing the rsp (stack pointer)
I noticed it was always decrementing further and further. This means the
stack is growing.

So I swapped out this recursion with a jump which does not put the
return address on the stack. And the stack pointer stayed at a fixed value.