Fix the bug!! Thanks @MilkeeyCat
recvfrom writes to the size argument. And it expects a pointer to an integer.
But a C styled integer so thats 4 bytes. While `db` or `equ` only are 1 byte.
Changing it to `dw` fixes that. Also ``db 128, 0, 0, 0`` would work.

But if it is only one byte it writes into the next structure which was
the server address struct. Thats why that one would get corrupted.